The latest Piglet commit fixes a number of issues: Piglet can now pick her own memory values to test, completely on her own, and there's no longer a max count for that (although she usually only has around 20-30 before I get bored and shut her off).
Piglet is now capable of looking at previous results and picking key combos that give her the best probabilities for success, but after about 30-40 frames, she'll run out of memory and die. I'm pretty sure that's just a bug somewhere though, and not a significant performance concern.

And I made a final decision to actually change Piglet's gender back to female. This actually turned out to be more of a difficult decision then you would expect, for fairly stupid reasons: back when I first started writing the AI, I initially intended Piglet to be a girl. When I changed her name though, I named her after a male character.

That wouldn't normally be as much of an issue for me- Piglet's a pretty gender neutral name to begin with. But apparently it's very common for people to get that gender wrong when referencing the books, and I've been going back and forth trying to decide whether or not keeping the original gender I planned would be worth dealing with all of the comments I'm going to get at launch saying, "Hey, did you know that Piglet is actually a boy in the books? Did you know that when you named her?"

Yes. I know and knew. You have literally no idea of the incredible extent that I know those books and those characters.

In any case, real progress is being made again. Piglet is still pretty stupid, but she has a strong basis to move forward, and it's encouraging to see the computer start making its own decisions instead of just watch me, even if (at the moment) those decisions are just as simple as, "I'd like to hold down right forever now." Things will only get better from here.

Signed-off-by: Daniel Shumway <shumway.danny@gmail.com>